packages/chelonia-server-v2/pyproject.toml
[tool.poetry]
name = "chelonia-server-v2"
version = "0.1.0"
description = ""
authors = ["hallblazzar <hallblazzar@gmail.com>"]
readme = "README.md"
packages = [{include = "chelonia_server_v2"}]

[tool.poetry.dependencies]
python = "^3.11"
django = {extras = ["bcrypt"], version = "^5.0.3"}
djangorestframework = "^3.14.0"
django-filter = "^23.5"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
dj-rest-auth = {extras = ["with-social"], version = "^5.0.2"}
pytest-django = "^4.8.0"
boto3 = "^1.34.62"
pycountry = "^23.12.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
